Updated openstack/openstack Project: openstack-infra/zuul  c908abf822a85a461c9563e17a0c8eca69c377b0
Support pagination in simplequery for Gerrit >=2.9

Starting with Gerrit 2.9 the pagination mechanism changed to using
offsets instead of sortkeys. For details please check the release
notes of 2.9:

https://gerrit-documentation.storage.googleapis.com/ReleaseNotes/ReleaseNotes-2.9.html#_rest_api.

In order to support both versions it is necessary to check for the
presence of moreChanges in the status line and then process the next
chunks either with the old or the new method.

Change-Id: Ieb54052debd07d86db66ef0f1d7ce1bdb42964b4