Updated openstack/openstack Project: openstack/nova  6f147d0b4ad1467d20ae7860e3d95e5cc74bcd87
Aggregate Extra Specs Filter should return if extra_specs is empty

Aggregate Extra Specs Filter should return if extra_specs is
empty. Aggregate Extra Specs Filter checks if extra specs attribute
is present as part of filter_properties; if it is not the filter
doesn't proceed further. But the filter tries to proceed further even
if the extra_specs attribute is present but its empty which should
not be done. This patch tries to fix the scenario where 'extra_specs'
is empty.

Change-Id: I23de8f5f45416c81e36e19bafb41e8c66c2ebd5c
Closes-Bug: #1512511