Updated openstack/openstack Project: openstack/nova  56dc61d7526a8f78813b902f7a76a9372a4b09b2
Fixes Python 3 str issue in ConfigDrive creation

ConfigDrive creation fails on Python 3 when writing metadata to file
due to the fact that in Python 3 only bytes can be written to a file.

This patch fixes this issue by replacing jsonutils.dumps with
jsonutils.dump_as_bytes and thus ensuring that all data that is to
be written is bytes and not str.

Change-Id: Ie84f8badb185295fe019d80f8036aeb941709342
Closes-Bug: #1502583