Updated openstack/openstack Project: openstack/manila  0c549ead18f22c5433135795da8752f51f1e3a3a
Refactor HP 3PAR share driver to now be HPE

This patch refactors the HP share driver to now be
the HPE share driver.

This is being done because the company responsible for
this driver is now Hewlett Packard Enterprise (HPE).

The driver is now located in the manila/share/drivers/hpe
folder.

DocImpact
Implements: blueprint rebrand-hp-3par-share-driver
Change-Id: I36efc04fa6da088ccb06070d8cd4fe29269cd7ca