Updated openstack/openstack Project: openstack/nova  3700c3c83a1ad7d803712eb88d107a051945f0b5
Remove IN-predicate warnings

The unit test logs currently contain 9 of the following warnings. Short
circuit and return an empty list in these cases rather than executing
the query needlessly.

    SAWarning: The IN-predicate on "<table>.<column>" was invoked with
an empty sequence. This results in a contradiction, which nonetheless
can be expensive to evaluate.  Consider alternative strategies for
improved performance.

Change-Id: I19460ed4e03a0db66186a4c231d315c2396b7eca