Updated openstack/openstack Project: openstack/ceilometer  f38091f2770c10e29cf2b2ef7683b1b069c6e540
Move the content of ReleaseNotes to README.rst

For now, we have added reno for release notes management, and a folder
"releasenotes" was added in ceilometer. Because there is a file named
"ReleaseNotes" to provide release note location and ceilometer home, it
will broken some git operations in case insensitive OS (e.g. Windows).
This change simply move the content of ReleaseNotes to README.rst to avoid
the problem.

Change-Id: Id883bef77eb74c798c8554d61db55e66c8d5e049