Updated openstack/openstack Project: openstack-infra/project-config  9857585dc4651e1fde46c6ea8693df31d5019d40
Don't run gate-horizon-dsvm-integration on stable branches

This job is non-voting and the Horizon team is working to get it voting
on master but it still looks to be unstable. Given it's non-voting and
not yet stable in master, let's not waste test resources running an
unreliable non-voting job on stable branches.

Change-Id: I284d9f597839801480f2ce0d04bf7d41d0281a45