Updated openstack/openstack Project: openstack/tripleo-heat-templates  37f712d07fec7341c0ca42cb175d370a20736a9f
Verify galera is sync'd in yum_update.sh

When the cluster is brought back online after a yum update in
yum_update.sh, we should verify that galera is fully sync'd before
moving on. This ensures the sync is complete before moving on to update
any other nodes in the cluster.

Change-Id: Ie8fc2c5d5214deacea94ca658ac75359b318ced1