Updated openstack/openstack Project: openstack/taskflow  c22bfc0ac39fbad5c41021a3a9a028933964a1fb
Update docstrings on entity type

Describe the instance variables and there types with
docstrings and comments and such, so that it is more
known what this type is for.

Also removes docstring saying this is only for jobs
and jobboards since it is really a more generic entity
class and has no direct tie to jobboards.

Change-Id: I98adb161adcb4ce096c1ee0e6c3377eb71383a90