Updated openstack/openstack Project: openstack/neutron  0b7d947db217a9753199f406318886cd0ea6eac0
Remove SysV init script for neutron-server

The script has some issues, f.e. it does not read any plugin
configuration file.

I don't believe anyone relies on it these days. The only distro I could
think of that could use it is Debian, but they have their own version of
the script maintained independently.

So it seems easier to remove it not to mislead users.

Change-Id: Ice622402a6261cfed8a8b042d732f39e25aff77c