Updated openstack/openstack Project: openstack-infra/devstack-gate  dfde77960d89e00eb780d0909b94230812a2a5a2
enable ebtables workaround in devstack

One of the biggest bugs in the gate right now is racing with libvirt
on ebtables for guests. This is because ebtables and the interfaces on
the kernel are not multi process safe. If you run ebtables twice
simultaneously it can bounce off the kernel or even possibly corrupt
the ebtables state.

This was addressed by adding a --concurrent flag in ebtables, however
every program calling ebtables must opt into using that flag. And if
anyone doesn't, the race remains. Nova is getting this support now,
libvirt gained it in 1.2.11, after metacloud discovered this
issue. However libvirt 1.2.11 is newer than we have in the gate.

We have an alternative workaround in devstack which replaces
/sbin/ebtables with a shell script that flocks and calls the real
ebtables behind the scene. This will work to stop bleeding out on this
bug and creating recheck madness.

Because this moves a file in /sbin, it's an opt in flag. As after a
package upgrade on a devstack it will be undone. However for the
single use case in devstack-gate it's fine, and should massively
decrease the race failures we find in the gate.

Depends-On: Ic6fa847eba34c21593b9df86a1c2c179534d0ba5

Change-Id: Ib48fc921c282ec7dad4a9e18460f2a064faae182
Related-Bug: #1501558