Updated openstack/openstack Project: openstack-infra/system-config  46ebbe3b10c1abe7f79e61960daa2dede88c54ac
Add puppetdb_file report processor

tl;dr This report processor consumes a puppet ruby report object and
writes a json wire-format report to disk in the puppet reportdir.

The report format produced by puppet does not match the wire format for
reports that are accepted by puppetdb so this report processor turns the
ruby report object produced by puppet into a json object in the format
that puppetdb expects to receive. The code for doing this already exists
in the puppetdb-terminus report processor, except that code then expects
to directly communicate with the puppetdb service to submit the report
and is not resuable. (I submitted
https://tickets.puppetlabs.com/browse/PDB-2150 to document this feature
request).

The masterless nodes in openstack-infra are not allowed to directly
communicate with the puppetdb service and so must have a custom report
processor that writes the json report to disk.

To enable this report processor, puppet must be configured with both
`report = true` (the default) and `reports = puppetdb_file`

Change-Id: I3b80a8012f2d6dc231c4d7701633d428a28b0006