Updated openstack/openstack Project: openstack-infra/nodepool  65719ca1c11e49e35e5f28f3158f3931485a11d0
Revert "Allow to delete nodes with missing target or provider"

This reverts commit 63397f1fe5172b1da876a9ede3cfbfe1218da077.

This causes nodes to leak.  Nodepool will delete nodes from the
database, but they still exist in the cloud.  The way to remove
a provider is to set max servers to -1, wait for it to cleanup,
then remove the provider.

Yes, if you don't follow that, there will be errors, but that's
because there are actually errors that nodepool can't recover
from.

Change-Id: Ie276e658e954473369402ef4aca98fe2b842b0fc