Updated openstack/openstack Project: openstack-infra/system-config  89ad2484c8e137ed4c6baaaa636b160590fb4d19
Pin puppetlabs-haproxy to commit containing 1.8 fix

There is a bug with ruby 1.8 and puppetlabs-haproxy that has been worked
around upstream. It manifests as an inability to compile the catalog
with haproxy on centos6.

Pin this to the commit with that fix applied until that commit is
released. Also, getting off of centos6 everywhere will make this problem
go away for real - as ruby 1.8 is EOL and not supported upstream.

Change-Id: I76d60b1e57553e144b8f5a7062e0061cc52067d7