Updated openstack/openstack Project: openstack/cinder  6e872f2758e1dc9680da6fa646e04a6d782ee753
Refactor HP LeftHand driver to now be HPE

This patch refactors the HP LeftHand driver to be
HPE LeftHand.

This is being done because the company responsible for
this driver is now Hewlett Packard Enterprise (HPE).

The driver is now located in the cinder/volume/drivers/hpe
folder.

DocImpact
Implements: blueprint rebrand-hp-lefthand-driver
Change-Id: I42eb5b3a51d547e45338a4964f31e0aca2ce43d8