Updated openstack/openstack Project: openstack-infra/irc-meetings  be837d9b18137efc190142579d6bae571ad19dbe
Removed No longer attended OpsMeeting

The Ops Community is trying to collapse our main meeting time into one
which, for the time being, is the Tools Working Group. This review
removes the "Team Meeting" because it is causing confusion.

I have confermation from IRC handles: mdorman, balajin, klindgern,
and raginbajin that this is ok.

Change-Id: I62a65b015429197d50ac54b2107a945a59ce7508