Updated openstack/openstack Project: openstack/cinder  b981262064cc27323fbd1faa0b231c7a4e427aad
Additional VAG support for SolidFire

This patch removes volumes from volume access groups, and potentially volume
access groups, when a connection is terminated. When the previously added
'sf_enable_vag' option is enabled and a volume connection is terminated, if the
volume was part of a VAG then the volume is removed from the VAG. If the volume
is the only remaining volume in the VAG then the VAG is also removed.

Change-Id: Ib508a67f012f6f186abba0d381ab104edc163fa9