Updated openstack/openstack Project: openstack/cinder  ae84ce085feaf9d021412c08ecd7d6a0a99d1e0f
ScaleIO driver: update_migrated_volume

This patch implements the update_migrated_volume function.
Without this change, when the migrate finishes, the original
volume will be deleted but the name on the new volume will not
be changed to the old id even though the Cinder database will
be updated. The volume name is a base64() encoding of the Cinder
volume id so any calls after the migrate fail since the lookup
fails.

Also changed 'SCALEIO' to connector.SCALEIO now that os-brick 0.4.0
is released.

Closes-Bug: #1512387
Change-Id: I70fe4bd651b0ac87b9208efb5918a4cdaf6f06b4