Updated openstack/openstack Project: openstack/keystone  e57b77cd1174b1c81a5945e07eddf131c5d3a420
Move oauth1 extension into core

Remove oauth1 as an extension and move it to a core resource.
For now we leave the database migrations in the extension directory
until we have a general policy for merging these into core.

DocImpact: update keystone-paste and remove oauth1 from pipeline

Change-Id: I0ed1ec44d42c3b379a5c2a40e3e6298842dfc01d
Implements: bp move-extensions