Updated openstack/openstack Project: openstack/keystone  4aa5fb4d8eac3d9e0db2cde956f4e58b571eb310
Move revoke extension into core

Remove revoke as an extension and move it to a core resource.
For now we leave the database migrations in the extension directory
until we have a general policy for merging these into core.

DocImpact: update keystone-paste and remove revoke from pipeline

Change-Id: I2d6f425a508b7acb4b4d079e4387f25bf7555683
Implements: bp move-extensions