Updated openstack/openstack Project: openstack/ironic-python-agent  a2a71105c4a758ddbfa206255b4bf320fc407af7
Refactor the image download and checksum computation bits

Prior to this patch downloading and computing the checksum of the image
were done in different stages, after the download the file would need
to be re-read and the checksum was computed. This patch is changing it
by creating a ImageDownload class which computes the checksum at the same
time the image is being downloaded, this saves time and also make the
code more portable.

Related-Bug: #1505685
Change-Id: I71f9f2bd9a62a6a6cc474d0ae519591cea6381d6