Updated openstack/openstack Project: openstack-infra/subunit2sql  ee1c388a7f711916ed2775b7ab2d99bf017eb343
Change uuid columns to bigint

This should lead to a smaller and much faster database. In addition
as part of the migration FKs get removed which will also have
some performance improvements.

Note, that FKs are left in the models on purpose. Some of the
sqlalchemy statements in the DB api use implicit joins and we
need the FK relationship in the model to allow those to work. This
will all be fixed in a follow up commit.

Co-Authored-By: Matthew Treinish <mtreinish@kortar.org>

Change-Id: Ie22ed9ef2ae0ad4fc0491ff93e77bc7c506c9cb4