Updated openstack/openstack Project: openstack/nova  ad94e5b0f4149b14267a64dffd2f7264b40ba394
docs: add the scheduler evolution plans

The scheduler evolution has been a priority for both kilo and liberty.
This document attempts to describe the general direction of that effort.

The goal for this work is to make it easier for others to help with this
effort, and make sure we agree the current direction.

Change-Id: I8746362b6a596690de21de34599462a6e27a92da