Updated openstack/openstack Project: openstack/python-keystoneclient  13bb2f74b0d65c1fef30f77d710d56e51e5f7841
Swap the order of username deprecation

The attempt at a move to user-name is an exercise in churn, and is
filling everyone's logs with admonitions to change the name of their
variables - which does not work if they do. Swap this, effectively
reverting the attempt at a move. user-name will continue to work on
the off chance anyone started consuming that path, which is unlikely
because none of the consuming programs expose that as an actual option.

Closes-Bug: 1498247

Change-Id: I62d991fda1df63c9cbabfde2f6836bc031f5147c