Updated openstack/openstack Project: openstack/governance  81d9bb950b3362b04fcd46f6c98f3b8f5b8608d3
Added JavaScript to Common Testing Interface

This patch expands the CTI documentation to support multiple
languages. It creates a single page that describes overall CTI goals,
and then links to documents that go into specifics on each language
supported in OpenStack.

The Javascript section is new, and includes documentation for the
current state of the Javascript CTI. It also explicitly calls out
areas where work needs to be done, in order to encourage
discussion and contributions.

Change-Id: I971b9271443e4fd227b4a6b347dfc385b67eb3e4
Depends-On: I00186c9af441c1778a22379634cdb2f918324bd8