Updated openstack/openstack Project: openstack-infra/subunit2sql  6e461277adf367c120454b64d0a40d6af65e2e90
Use the test_id column as the test identifier

This commit changes the failures and run_time graph commands in
subunit2sql-graph to take in a test_id instead of a test.id. This
is much more user friendly because it's nominally human readable.
We also just landed a large database migration that changes the
test.id column from a uuid string to an integer. This will break
users of the script and not give them a migration path. So instead of
having people rely on the internal identifier they should just use
the name.

Change-Id: I949a62e8beedd3b76d9d4673e658523f1a75253d