Updated openstack/openstack Project: openstack/python-openstackclient  4776bf672b6359e9d157e9e9ca871402d62e58c0
Enable FakeResource to fake methods.

Use MagicMock to fake a method in FakeResource. A new function:

    add_method(name, return_value)

is added to FakeResource. The caller specifies method @name and @return_value,
the function will add an attribute with @name, which is a callable MagicMock
object whose return value is @return_value.

When user access the attribute with a (), @return_value will be returned by
MagicMock, which looks like a function call.

Change-Id: I12eb876cbebab064773df7b5dd612de69bbf3f01
Implements: blueprint osc-unit-test-framework-improvement