Updated openstack/openstack Project: openstack-infra/puppet-log_processor  b2190b1a425db0cf61ce3c3df5fd5443aa1aee0f
Add a node_region field to the job metadata

The node region can be figured out from the build_node very easily and
having a discrete field will make filtering to a single region much
simpler. This commit adds a new metadata field 'node_region' which is
the cloud region that the build_node ran in.

Change-Id: I06bbb62d21871ee61dbfb911143efff376992b98