Updated openstack/openstack Project: openstack/openstack-manuals  9a45431e83ba67c815f9141f1adec4a0caf7a50e
Refactoring HP LeftHand Cinder Driver to be HPE

The HP LeftHand Cinder driver was recently refactored to be HPE.

The same refactoring is being done to the documentation with
this patch. It is being done because the company managing the HP
LeftHand Cinder driver is now Hewlett Packard Enterprise (HPE).

Closes-Bug: #1517648
Change-Id: Ibe4d4ff73591348fcf5997984e89d0a895172f6b