Updated openstack/openstack Project: openstack/openstack-doc-tools  1d0851cd74812ab83bd037b92df5913e22071ab3
Fix for getting glance command list

Usual command uses xxxx --help for getting command list.
but glance case, it need to use help not --help.
For this purpose, adds if branch.
This patch is needed for complete/bash-completion switch.

glance is not viewed "Subcommands" list at header
  glance (Subcommands does not exist)
  http://docs.openstack.org/cli-reference/content/glanceclient_commands.html
  cinder (Subcommands exist)
  http://docs.openstack.org/cli-reference/content/cinderclient_commands.html
For this purpose, add glance switch for "Subcommands".

Change-Id: I33251e5ada49c6e2228d5d08aa5f2c5fd86da8bb
Closes-Bug: #1517756