Updated openstack/openstack Project: openstack/tripleo-heat-templates  338a2bcfb30aa4aa1660e4eadf3d52c2e09a23c3
Update docker compute environment to use json config

In liberty, Kolla copies around files and runs the service given
a specified command, by reading a json file.

This will update the existing work to follow that template by
creating a json file for each of the services and pushing it
into the containers.

Change-Id: I5085d1896ea965fd8854765b055068a5ad30bcfd
Co-Authored-By: Jeff Peeler <jpeeler@redhat.com>