Updated openstack/openstack Project: openstack/oslo.config  9dd719f83aa888efee7f9bcbc8799e86b2ec8504
Add a bandit target to tox.ini

This requires some changes in the Python code, because bandit considers:

    try:
        <some code>
    except <some error>:
        pass

as an issue, since it might be an involuntarily unhandled error.

So this pattern is changed to something else when easily doable. Otherwise,
"nosec" is added when the code is obviously safe.

Change-Id: Ic99f29b944892fb78814f5de515e6ed699979bc1