Updated openstack/openstack Project: openstack-infra/project-config  da8c04f38378c28d6663b63da9ba811cb79f53fb
Fix job definition for ceph-src jobs

The job definition for ceph-src was doing some weird stuff. First it
was setting both TEMPEST_FULL and a DEVSTACK_GATE_REGEX, those 2
are contradictory, and DEVSTACK_GATE_REGEX will take precedence.
So, this patch removes the TEMPEST_FULL variable. The second issue
is that regex used was wrong, and it was running more tests then was
intended, they also need to skip slow tests. Slow tests should not be
run as part of a gating job, because as the tag name implies, they are
too slow. (periodic and experimental are fine though) So this patch
corrects the regex to properly exclude slow and the test which exposes
where the ceph drivers break.

Change-Id: I4f70a3785ea4aeb38c3ebd8fd6a74ed3d035fa1a